requirements.txt
streamlit==1.28.1
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
requests==2.31.0
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
huggingface-hub==0.17.3
accelerate==0.24.1